# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Academia Fight Club API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Confirm a user's email using the token sent at registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend the verification email to the logged-in user
      responses:
        "200":
          description: Verification email resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: sessionId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: modality
          in: query
          schema:
            type: string
            enum: [thai, jiu]
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create attendance record (mark student present)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Remove an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /attendance/bulk:
    post:
      operationId: bulkAttendance
      tags: [attendance]
      summary: Mark attendance for many students in a single modality
      description: |
        Finds (or creates) today's session for the given modality and inserts an
        attendance record for each student if one does not already exist.
        Server-side dedupe prevents duplicate presence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAttendanceInput"
      responses:
        "200":
          description: Bulk attendance result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAttendanceResult"
        "400":
          description: Bad request
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, teacher]
        unit:
          type: string
          enum: [matriz, panobianco, upfitness]
          nullable: true
        phone:
          type: string
        birthDate:
          type: string
          nullable: true
        paymentDay:
          type: integer
          nullable: true
          minimum: 1
          maximum: 31
        modalityThai:
          type: boolean
          nullable: true
        modalityJiu:
          type: boolean
          nullable: true
        bollacha:
          type: boolean
          nullable: true
        thaiGrade:
          type: string
          nullable: true
        thaiGradeColor:
          type: string
          nullable: true
        jiuGrade:
          type: string
          nullable: true
        jiuGradeColor:
          type: string
          nullable: true
        jiuDegree:
          type: integer
          nullable: true
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    VerifyEmailInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, unit, emailVerified, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        unit:
          type: string
          enum: [matriz, panobianco, upfitness]
        emailVerified:
          type: boolean
        phone:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        paymentDay:
          type: ["integer", "null"]
        modalityThai:
          type: ["boolean", "null"]
        modalityJiu:
          type: ["boolean", "null"]
        thaiGrade:
          type: ["string", "null"]
        thaiGradeColor:
          type: ["string", "null"]
        jiuGrade:
          type: ["string", "null"]
        jiuGradeColor:
          type: ["string", "null"]
        jiuDegree:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    AttendanceRecord:
      type: object
      required: [id, sessionId, studentId, studentName, modality, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        studentPhotoUrl:
          type: ["string", "null"]
        modality:
          type: string
          enum: [thai, jiu]
        postTrainingPhotoUrl:
          type: ["string", "null"]
        faceRecognized:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AttendanceInput:
      type: object
      required: [sessionId, studentId]
      properties:
        sessionId:
          type: integer
        studentId:
          type: integer
        postTrainingPhotoUrl:
          type: string
        faceRecognized:
          type: boolean
...
    BulkAttendanceInput:
      type: object
      required: [teacherId, modality, students]
      properties:
        teacherId:
          type: integer
        modality:
          type: string
          enum: [thai, jiu]
          description: Single modality chosen by the teacher for this team photo; applied to every student in the list.
        photoUrl:
          type: string
        students:
          type: array
          items:
            type: integer
          description: Student user ids to mark present.
    BulkAttendanceResult:
      type: object
      required: [created, skipped]
      properties:
        created:
          type: integer
        skipped:
          type: integer