# lib/api-zod/src/generated/types/bulkAttendanceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academia Fight Club API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkAttendanceInputModality } from "./bulkAttendanceInputModality";

export interface BulkAttendanceInput {
  teacherId: number;
  /** Single modality chosen by the teacher for this team photo; applied to every student in the list. */
  modality: BulkAttendanceInputModality;
  photoUrl?: string;
  /** Student user ids to mark present. */
  students: number[];
}