# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Academia Fight Club API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserUnit } from "./userUnit";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  unit: UserUnit;
  emailVerified: boolean;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  paymentDay?: number | null;
  /** @nullable */
  modalityThai?: boolean | null;
  /** @nullable */
  modalityJiu?: boolean | null;
  /** @nullable */
  thaiGrade?: string | null;
  /** @nullable */
  thaiGradeColor?: string | null;
  /** @nullable */
  jiuGrade?: string | null;
  /** @nullable */
  jiuGradeColor?: string | null;
  /** @nullable */
  jiuDegree?: number | null;
  createdAt: Date;
}